Adds support for version 204 of the Sift Science API.
Version 204 of the Sift Science API is now called by default -- this
is an incompatible change, as some fields are no longer returned in
the v204 API (unless they are explicitly requested).

Also cleans up all Client methods (and the constructor) using an Array
argument for optional paramaters.  This is an incompatible change.

Also adds support for the Workflow Status API, User Decisions API, and
Order Decisions API.